requirements.txt
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
Pillow==12.2.0
requests==2.33.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
whitenoise==6.8.2
